+++ lib/generators/ductwork/update/templates/db/add_crash_count_to_ductwork_advancements.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddCrashCountToDuctworkAdvancements < Ductwork::Migration
+  def change
+    add_column :ductwork_advancements, :crash_count, :integer
+  end
+end