+++ lib/generators/ductwork/update/update_generator.rb
@@ -82,6 +82,11 @@ module Ductwork
                            "db/migrate/add_crash_count_to_ductwork_executions.rb"
       end
 
+      if Ductwork::Advancement.column_names.exclude?("crash_count")
+        migration_template "db/add_crash_count_to_ductwork_advancements.rb",
+                           "db/migrate/add_crash_count_to_ductwork_advancements.rb"
+      end
+
       if Ductwork::Execution.column_names.exclude?("process_id")
         migration_template "db/add_process_id_to_ductwork_executions.rb",
                            "db/migrate/add_process_id_to_ductwork_executions.rb"